added algorithm output visualization
I considered an approach of the problem so that the individuals with less connections could form by themselves a community.
I though that this approach would make sense when considering their limited knowledge of the surroundings, as they form a line, compared to the "hub" formations we can see in the larger communities. We form communities based on the level of interconnections and the immediate impact of propagation of a signal through the network.
Therefore, the "line" graphs would have little to no effect when in their community a signal is produced, rather that the "hubs" or cliques where any news spread so fast.
Another approach could be that the cliques could be considered as end devices of a network and the "lines" are the wires. As we know from the real world, inside a machine all the signals are transferred immediatelly and wired connections have multiple issues such as less data transfer speed. Therefore, "lines" and "hubs" are considered here different communities.